+++ app/src/components/FeaturedRow.tsx
@@ -1,7 +1,7 @@
 import type { HydratedObjective } from "@/types/objectives";
 
 import { Separator } from "@/components/ui/separator";
-import { PathCard } from "@/components/cards/PathCard";
+import { PathCard } from "@/components/cards/ObjectiveCard";
 
 type PropTypes = {
   paths: Array<HydratedObjective>;
