+++ app/src/routes/browse.tsx
@@ -6,7 +6,7 @@ import type { HydratedObjective } from "@/types/objectives";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Separator } from "@/components/ui/separator";
-import { PathCard } from "@/components/cards/PathCard";
+import { PathCard } from "@/components/cards/ObjectiveCard";
 
 import { hydrateObjectives } from "@/lib/getData";
